Updates from Lauren; delete zip
from @lkmancewicz-erdc commits in https://github.com/LimnoTech/modflow-modeling
- Increased temporal resolution for mf6
- Added plot to view order of grid cell IDs
Also
- Added capability to modify well flowrate
Also, deleted the zip file of mf6 inputs, because I decided (in previous commit) that it would be better to expose those files to useers and git-track any changes we decide to make on them (i.e. if we refine the grid).